refactor: [FE] Main 컴포넌트 리팩토링
- 코드 유지보수 및 렌더링 최적화를 위해 Main 컴포넌트를 하위
  컴포넌트들의 조합으로 분리
