+++ frontend/src/components/Main/Main.scss
@@ -12,55 +12,6 @@
     border: 1px solid #ffffff;
     border-radius: 5px;
     background-color: #000000;
-
-    aside {
-      width: 15%;
-    }
-
-    .audi-main-menu-container{
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      width: 5%;
-      height: 100%;
-      margin: 0 15px;
-
-      .audi-main-track-option-area{
-        padding-top: 1.8rem;
-      }
-    }
-
-    .audi-main-audio-track-container {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      position: relative;
-      width: 80%;
-      height: 100%;
-      background-color: rgba(0, 0, 0, 0);
-      overflow: hidden;
-
-      .audi-main-audio-track-scroll-area{
-        overflow-x: scroll;
-        -ms-overflow-style: none; 
-        scrollbar-width: none;
-
-        &::-webkit-scrollbar {
-            display: none; 
-        }
-      }
-
-      .audi-main-audio-track-container-event-zone {
-        position: absolute;
-        width: 100%;
-        height: 95%;
-
-        top: 2rem;
-        left: 0;
-
-        z-index: 90;
-      }
-    }
   }
 
   .cursor-change {
@@ -71,20 +22,4 @@
   }
 }
 
-@media screen and (max-width: 1680px) {
-  .audi-main-container {
-    .audi-main-content {
-      aside {
-        width: 22%;
-      }
-  
-      .audi-main-menu-container{
-        width: 8%;
-      }
 
-      .audi-main-audio-track-container {
-        width: 70%;
-      }
-    }
-  }
-}
