+++ frontend/src/components/Main/MainAsideContent/MainAsideContent.scss
@@ -1,0 +1,9 @@
+audi-main-aside-content{
+    width: 15%;
+    height: 100%;
+
+    aside {
+        width: 100%;
+        height: 100%;
+    }
+}
