+++ frontend/src/components/Main/MainMiddleContent/MainMiddleContent.scss
@@ -1,0 +1,18 @@
+audi-main-middle-content{
+    width: 8%;
+
+    .audi-main-menu-container{
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        align-items: center;
+        width: 100%;
+        height: 100%;
+        padding: 0 12%;
+
+        .audi-main-track-option-area{
+            width: 100%;
+            padding-top: 1.8rem;
+        }
+    }
+}
