+++ frontend/src/components/Main/MainRightContent/MainRightContent.scss
@@ -1,0 +1,35 @@
+audi-main-right-content{
+    width: 77%;
+
+    .audi-main-audio-track-container {
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        position: relative;
+        width: 100%;
+        height: 100%;
+        background-color: rgba(0, 0, 0, 0);
+        overflow: hidden;
+    
+        .audi-main-audio-track-scroll-area{
+          overflow-x: scroll;
+          -ms-overflow-style: none; 
+          scrollbar-width: none;
+    
+          &::-webkit-scrollbar {
+              display: none; 
+          }
+        }
+    
+        .audi-main-audio-track-container-event-zone {
+          position: absolute;
+          width: 100%;
+          height: 95%;
+    
+          top: 2rem;
+          left: 0;
+    
+          z-index: 90;
+        }
+    }
+}
