+++ frontend/src/components/Main/index.js
@@ -1,0 +1,4 @@
+import './MainRightContent/MainRightContent';
+import './MainMiddleContent/MainMiddleContent';
+import './MainAsideContent/MainAsideContent';
+import './Main';
